Crate linera_base

This module provides a common set of types and library functions that are shared between the Linera protocol (compiled from Rust to native code) and Linera applications (compiled from Rust to Wasm).

Modules

abi

This module defines the notion of Application Binary Interface (ABI) for Linera applications across Wasm and native architectures.

command

Command functionality used for spanning child processes.

crypto

Define the cryptographic primitives used by the Linera protocol.

data_types

Core data-types used in the Linera protocol.

identifiers

Core identifiers used by the Linera protocol.

ownership

Structures defining the set of owners and super owners, as well as the consensus round types and timeouts for chains.

prometheus_util

This module defines util functions for interacting with Prometheus (logging metrics, etc)

sync

Provide Lazy struct on top of std. Based on once_cell::sync::Lazy.

Macros

bcs_scalar

Defines a GraphQL scalar type using the hex-representation of the value’s BCS-serialized form.

doc_scalar

Defines a GraphQL scalar with a description string.

ensure

A macro for asserting that a condition is true, returning an error if it is not.

Enums

BcsHexParseError

An error trying to parse the hex-digits of a BCS-encoded value.

Functions

hex_debug

Formats a byte sequence as a hexadecimal string, and elides bytes in the middle if it is longer than 32 bytes.