Struct linera_base::committee::Committee

pub struct Committee {
    pub validators: BTreeMap<ValidatorName, ValidatorState>,
    pub total_votes: u64,
    pub quorum_threshold: u64,
    pub validity_threshold: u64,
}

A set of validators (identified by their public keys) and their voting rights.

Fields
validators: BTreeMap<ValidatorName, ValidatorState>

The validators in the committee.

total_votes: u64

The sum of all voting rights.

quorum_threshold: u64

The threshold to form a quorum.

validity_threshold: u64

The threshold to prove the validity of a statement.

Implementations

impl Committee

pub fn new(validators: BTreeMap<ValidatorName, ValidatorState>) -> Self

pub fn make_simple(keys: Vec<ValidatorName>) -> Self

pub fn weight(&self, author: &ValidatorName) -> u64

pub fn network_address(&self, author: &ValidatorName) -> Option<&str>

pub fn quorum_threshold(&self) -> u64

pub fn validity_threshold(&self) -> u64

pub fn get_lower_bound<V>(
&self,
threshold: u64,
values: Vec<(ValidatorName, V)>
) -> Vwhere
V: Default + Ord,

Find the highest value than is supported by a certain subset of validators.

pub fn get_strong_majority_lower_bound<V>(
&self,
values: Vec<(ValidatorName, V)>
) -> Vwhere
V: Default + Ord,

Find the highest value than is supported by a quorum of validators.

pub fn get_validity_lower_bound<V>(&self, values: Vec<(ValidatorName, V)>) -> Vwhere
V: Default + Ord,

Find the highest value than is guaranteed to be supported by at least one honest validator.