Trait linera_base::crypto::BcsHashable

pub trait BcsHashable: Serialize + DeserializeOwned { }

Activate the blanket implementation of Hashable based on serde and BCS.

We use serde_name to extract a seed from the name of structs and enums.

We use BCS to generate canonical bytes suitable for hashing.