Module linera_base::identifiers

Core identifiers used by the Linera protocol.

Structs

Account

A system account.

ApplicationId

A unique identifier for a user application.

BlobId

A content-addressed blob ID i.e. the hash of the Blob.

BytecodeId

A unique identifier for an application bytecode.

ChainId

The unique identifier (UID) of a chain. This is currently computed as the hash value of a ChainDescription.

ChannelName

The name of a subscription channel.

MessageId

The index of a message in a chain.

Owner

The owner of a chain. This is currently the hash of the owner’s public key used to verify signatures.

Enums

AccountOwner

An account owner.

ChainDescription

How to create a chain.

Destination

The destination of a message, relative to a particular application.

GenericApplicationId

A unique identifier for an application.