linera_base
Module prometheus_util

This module defines utility functions for interacting with Prometheus (logging metrics, etc)

Structs

ActiveMeasurementGuard

A guard for an active latency measurement.

Traits

MeasureLatency

An extension trait for metrics that can be used to measure latencies.

Functions

exponential_bucket_interval

Construct the bucket interval exponentially starting from a value and an ending value.

exponential_bucket_latencies

Construct the latencies exponentially starting from 0.001 and ending at the maximum latency

linear_bucket_interval

Construct the bucket interval linearly starting from a value and an ending value.

linear_bucket_latencies

Construct the latencies linearly starting from 1 and ending at the maximum latency

register_histogram_vec

Wrapper around Prometheus register_histogram_vec! macro which also sets the linera namespace

register_int_counter_vec

Wrapper around Prometheus register_int_counter_vec! macro which also sets the linera namespace