linera_base
Module task

Abstractions over tasks that can be used natively or on the Web.

Structs

Blocking

A new task running in a different thread.

SendError

The type of errors that can result from sending a message to the spawned task. Error returned by the Sender.

Traits

Post

Types that can be explicitly sent to a new thread. This differs from Send in that we can provide an explicit post step (e.g. postMessage on the Web).

Functions

spawn

Spawns a new task, potentially on the current thread.

Type Aliases

BlockingFuture

The type of a future awaiting another thread.

InputReceiver

The stream of inputs available to the spawned task.

NoInput

A type that satisfies the send/receive bounds, but can never be sent or received.

NonBlockingFuture

The type of a future awaiting another task.