linera_base
Module time

Abstractions over time that can be used natively or on the Web.

Modules

timer

Utilities for tracking time.

Structs

Duration

A Duration type to represent a span of time, typically used for system timeouts.

Instant

A measurement of a monotonically nondecreasing clock. Opaque and useful only with Duration.

SystemTime

A measurement of the system clock, useful for talking to external entities like the file system or other processes.

SystemTimeError

An error returned from the duration_since and elapsed methods on SystemTime, used to learn how far in the opposite direction a system time lies.

TryFromFloatSecsError

An error which can be returned when converting a floating-point value of seconds into a Duration.

Constants

UNIX_EPOCH

An anchor in time which can be used to create new SystemTime instances or learn about where in time a SystemTime lies.