Crate linera_base

This module provides a common set of types and library functions that are shared between the Linera protocol (compiled from Rust to native code) and Linera applications (compiled from Rust to Wasm).

Modules

crypto

data_types

identifiers

Macros

ensure

A macro for asserting that a condition is true, returning an error if it is not.

Functions

hex_debug

Formats a byte sequence as a hexadecimal string.