linera_base

Module command

Command functionality used for spawning child processes.

Traits

CommandExt

Extension trait for tokio::process::Command.

Functions

binary_parent

Retrieves the path from the binary parent.

current_binary_parent

Obtains the current binary parent

parse_version_message

Obtains the version from the message.

resolve_binary

Attempts to resolve the path and test the version of the given binary against our package version.

resolve_binary_in_same_directory_as

Same as resolve_binary but gives the option to specify a binary path to use as reference. The path may be relative or absolute but it must point to a valid file on disk.