Module linera_base::crypto

Define the cryptographic primitives used by the Linera protocol.

Structs

CryptoHash

A Sha3-256 value.

KeyPair

A signature key-pair.

PublicKey

A signature public key.

Signature

A signature value.

TestString

A BCS-signable struct for testing.

Enums

CryptoError

Error type for cryptographic errors.

Traits

BcsHashable

Activate the blanket implementation of Hashable based on serde and BCS.

BcsSignable

Activate the blanket implementation of Signable based on serde and BCS.

CryptoRng

Wrapper around rand::CryptoRng and rand::RngCore.

HasTypeName

Something that we know how to hash and sign.

Hashable

Something that we know how to hash.