Module linera_base::data_types

Core data-types used in the Linera protocol.

Structs

Amount

A non-negative amount of tokens.

ApplicationPermissions

Permissions for applications on a chain.

BlockHeight

A block height to identify blocks in a chain.

Resources

Resources that an application may spend during the execution of transaction or an application call.

SendMessageRequest

A request to send a message.

Timestamp

A timestamp, in microseconds since the Unix epoch.

Enums

ArithmeticError

An error type for arithmetic errors.

ParseAmountError

Round

An identifier for successive attempts to decide a value in a consensus protocol.