linera_base

Module data_types

Core data-types used in the Linera protocol.

Structs

Amount

A non-negative amount of tokens.

AmountConversionError

Error converting from U256 to Amount. This can fail since Amount is a u128.

ApplicationDescription

Description of a user application.

ApplicationPermissions

Permissions for applications on a chain.

Blob

A blob of binary data, with its hash.

BlobContent

A blob of binary data.

BlockHeight

A block height to identify blocks in a chain.

Bytecode

A WebAssembly module’s bytecode.

ChainDescription

Initial chain configuration and chain origin.

CompressedBytecode

A compressed module bytecode (WebAssembly or EVM).

Epoch

A number identifying the configuration of the chain (aka the committee).

Event

An event recorded in a block.

InitialChainConfig

The initial configuration for a new chain.

NetworkDescription

A description of the current Linera network to be stored in every node’s database.

Resources

Resources that an application may spend during the execution of transaction or an application call.

SendMessageRequest

A request to send a message.

StreamUpdate

An update for a stream with new events.

TimeDelta

A duration in microseconds.

Timestamp

A timestamp, in microseconds since the Unix epoch.

Enums

ArithmeticError

An error type for arithmetic errors.

ChainOrigin

What created a chain.

DecompressionError

A type for errors happening during decompression.

OracleResponse

A record of a single oracle response.

ParseAmountError

Round

An identifier for successive attempts to decide a value in a consensus protocol.