Module linera_base::data_types

Core data-types used in the Linera protocol.

Structs

Amount

A non-negative amount of tokens.

ApplicationPermissions

Permissions for applications on a chain.

Blob

A blob of binary data, with its hash.

BlobContent

A blob of binary data.

BlockHeight

A block height to identify blocks in a chain.

Bytecode

A WebAssembly module’s bytecode.

CompressedBytecode

A compressed WebAssembly module’s bytecode.

Resources

Resources that an application may spend during the execution of transaction or an application call.

SendMessageRequest

A request to send a message.

TimeDelta

A duration in microseconds.

Timestamp

A timestamp, in microseconds since the Unix epoch.

UserApplicationDescription

Description of the necessary information to run a user application.

Enums

ArithmeticError

An error type for arithmetic errors.

DecompressionError

A type for errors happening during decompression.

OracleResponse

A record of a single oracle response.

ParseAmountError

Round

An identifier for successive attempts to decide a value in a consensus protocol.