Module linera_base::data_types

Structs

BlockHeight

A block height to identify blocks in a chain.

ChainId

The unique identifier (UID) of a chain. This is the hash value of a ChainDescription.

EffectId

The index of an effect in a chain.

Epoch

A number identifying the configuration of the chain (aka the committee).

Owner

The owner of a chain. This is currently the hash of the owner’s public key used to verify signatures.

RoundNumber

A number to identify successive attempts to decide a value in a consensus protocol.

Timestamp

A timestamp, in microseconds since the Unix epoch.

ValidatorName

The identity of a validator.

Enums

ArithmeticError

An error type for arithmetic errors.

ChainDescription

How to create a chain.