linera_base

Module identifiers

Core identifiers used by the Linera protocol.

Structs

Account

An account.

ApplicationId

A unique identifier for a user application from a blob.

BlobId

A content-addressed blob ID i.e. the hash of the BlobContent.

ChainId

The unique identifier (UID) of a chain. This is currently computed as the hash value of a ChainDescription.

DataBlobHash

Hash of a data blob.

EventId

An event identifier.

IndexAndEvent

The result of an events_from_index.

ModuleId

A unique identifier for a module.

StreamId

An event stream ID.

StreamName

The name of an event stream.

Enums

AccountOwner

An account owner.

BlobType

The type of the blob. Should be a 1:1 mapping of the types in Blob.

GenericApplicationId

A unique identifier for an application.