Module linera_base::identifiers

Structs

ApplicationId

A unique identifier for a user application.

BytecodeId

A unique identifier for an application bytecode.

ChainId

The unique identifier (UID) of a chain. This is currently computed as the hash value of a ChainDescription.

ChannelName

The name of a subscription channel.

MessageId

The index of a message in a chain.

Owner

The owner of a chain. This is currently the hash of the owner’s public key used to verify signatures.

SessionId

The identifier of a session.

Enums

ChainDescription

How to create a chain.

Destination

The destination of a message, relative to a particular application.