Trait linera_base::prometheus_util::MeasureLatency

pub trait MeasureLatency: Sized {
    // Required methods
    fn measure_latency(&self) -> ActiveMeasurementGuard<'_, Self>;
    fn finish_measurement(&self, milliseconds: f64);
}

An extension trait for metrics that can be used to measure latencies.

Required Methods

fn measure_latency(&self) -> ActiveMeasurementGuard<'_, Self>

Starts measuring the latency, finishing when the returned ActiveMeasurementGuard is dropped.

fn finish_measurement(&self, milliseconds: f64)

Updates the metric with measured latency in milliseconds.

Object Safety

This trait is not object safe.